src/word/requirements.txt
Flask==2.1.1
google-cloud-firestore==2.8.0
gunicorn
